+++ hubspot_tech_writing/util/common.py
@@ -1,4 +1,6 @@
 import logging
+import typing as t
+from pathlib import Path
 
 import colorlog
 from colorlog.escape_codes import escape_codes
@@ -14,3 +16,25 @@ def setup_logging(level=logging.INFO, verbose: bool = False):
     logging.basicConfig(format=log_format, level=level, handlers=[handler])
 
     logging.getLogger("urllib3.connectionpool").setLevel(level)
+
+
+class ContentTypeResolver:
+    MARKUP_SUFFIXES = [".md", ".rst"]
+    HTML_SUFFIXES = [".html", ".html5", ".htm"]
+    TEXT_SUFFIXES = MARKUP_SUFFIXES + HTML_SUFFIXES + [".txt"]
+
+    def __init__(self, name: t.Union[str, Path]):
+        self.name = name
+        self.suffix = Path(name).suffix
+
+    def is_markup(self):
+        return self.suffix in self.MARKUP_SUFFIXES
+
+    def is_html(self):
+        return self.suffix in self.HTML_SUFFIXES
+
+    def is_text(self):
+        return self.suffix in self.TEXT_SUFFIXES
+
+    def is_file(self):
+        return not self.is_text()