+++ tests/test_cli.py
@@ -1,3 +1,5 @@
+from unittest.mock import Mock
+
 import pytest
 from click.testing import CliRunner
 
@@ -57,7 +59,7 @@ def test_linkcheck_broken(caplog, markdownfile_minimal_broken_links):
     assert "[✖] https://foo.example.org/" in result.output
 
 
-def test_upload_no_access_token(caplog, markdownfile):
+def test_upload_no_access_token(markdownfile):
     runner = CliRunner()
 
     with pytest.raises(ValueError) as ex:
@@ -67,3 +69,25 @@ def test_upload_no_access_token(caplog, markdownfile):
             catch_exceptions=False,
         )
     assert ex.match("Communicating with the HubSpot API needs an access token")
+
+
+def test_delete_blogpost(mocker):
+    runner = CliRunner()
+    delete_blogpost: Mock = mocker.patch("hubspot_tech_writing.cli.delete_blogpost")
+    runner.invoke(
+        cli,
+        args="--debug delete post --id=138458225506 --access-token=foo",
+        catch_exceptions=False,
+    )
+    delete_blogpost.assert_called_once_with(access_token="foo", identifier="138458225506", name=None)  # noqa: S106
+
+
+def test_delete_file(mocker):
+    runner = CliRunner()
+    delete_file: Mock = mocker.patch("hubspot_tech_writing.cli.delete_file")
+    runner.invoke(
+        cli,
+        args="--debug delete file --id=138458225506 --access-token=foo",
+        catch_exceptions=False,
+    )
+    delete_file.assert_called_once_with(access_token="foo", identifier="138458225506", path=None)  # noqa: S106