+++ CHANGES.md
@@ -8,7 +8,9 @@
 - Add initial converter implementation
 - HTML: Add addons for custom code blocks, and header links
 - Add linkcheck subsystem
-- Add HubSpot API subsystem and "upload" feature
+- Add HubSpot "BlogPost" wrapper and "upload" feature
+- Add HubSpot "File" wrapper, with refactoring
+- Add "delete" operations for both "BlogPost", and "File" entities
 
 
 ## 2023/09/xx 0.0.0